added Sender class component of the Log mechnism
The Sender class is made to be inherited from so that the derived class
gets logging functionality. It uses the a reference to the Receiver
class which is instantiated from a singleton. There is only one receiver
instance. So I may change that later to provide the ability to use
several loggers at once.